add collectionSearchBar Component
