+++ components/BikeCard/index.jsx
@@ -1,4 +1,5 @@
 /* :::: BikeCard :::: */
+import Link from 'next/link';
 import BikeCardStyle from '../../styles/BikeCardStyle.module.css';
 
 
@@ -12,7 +13,9 @@ return (<section className={BikeCardStyle.card_wrap}>
     key={bike._id}
     className={BikeCardStyle.bikecard}
     >
+    <Link href={`./details/[slug].js`}>
     <button className={BikeCardStyle.detail_button}>&#187;</button>
+    </Link>
     <div className={BikeCardStyle.bikecard_fragment}>
         <p>Bike Title: {bike.bikeTitle}</p>
         <p>Bike Type: {bike.bikeType}</p>
