+++ components/CollectionSearchBar/index.jsx
@@ -1,0 +1,12 @@
+/* :::: CollectionSearchBar :::: */
+import styles from '@/styles/Header.module.css'
+
+
+export default function CollectionSearchBar(){
+
+
+return (<section className={BikeCardStyle.card_wrap}>
+    <ul>
+</ul>
+</section>)
+}
